Refactor login routes
