+++ spec/controllers/api/v1/auth/users_controller_spec.rb
@@ -2,74 +2,6 @@ require 'rails_helper'
 
 RSpec.describe Api::V1::Auth::UsersController, type: :controller do
   describe '#create' do
-    let!(:db_user) { create(:user) }
-
-    context 'When signing in with a valid user' do
-      let(:result) {
-        {
-          auth_token: 'token',
-          user: {
-            username: 'user'
-          }
-        }
-      }
-
-      before do
-        allow(JsonWebToken).to receive(:encode).and_return('token')
-        post :create, params: { login: 'user', password: 123456 }
-      end
-
-      it 'has a 200 status code' do
-        expect(response.status).to eq(200)
-      end
-
-      it 'returns json with payload' do
-        expect(response.body).to eq result.to_json
-      end
-    end
-
-    context 'When signing in with a invalid password' do
-      let(:result) {
-        {
-          errors: ['Invalid Username/Password']
-        }
-      }
-
-      before do
-        post :create, params: { login: 'user', password: 'invalid_password' }
-      end
-
-      it 'has a 401 status code' do
-        expect(response.status).to eq(401)
-      end
-
-      it 'returns json with errors' do
-        expect(response.body).to eq result.to_json
-      end
-    end
-
-    context 'When signing in with a nonexistent user' do
-      let(:result) {
-        {
-          errors: ['Invalid Username/Password']
-        }
-      }
-
-      before do
-        post :create, params: { login: 'invalid_user', password: 'invalid_password' }
-      end
-
-      it 'has a 401 status code' do
-        expect(response.status).to eq(401)
-      end
-
-      it 'returns json with errors' do
-        expect(response.body).to eq result.to_json
-      end
-    end
-  end
-
-  describe '#sign_up' do
     context 'When has no errors' do
       let(:result) {
         {
@@ -84,7 +16,7 @@ RSpec.describe Api::V1::Auth::UsersController, type: :controller do
         allow(AuthService).to receive(:sign_up!).and_return(create(:user))
         allow(JsonWebToken).to receive(:encode).and_return('token')
 
-        post :sign_up
+        post :create
       end
 
       it 'has a 200 status code' do
@@ -111,7 +43,7 @@ RSpec.describe Api::V1::Auth::UsersController, type: :controller do
         invalid_user = User.create()
         allow(AuthService).to receive(:sign_up!).and_raise(ActiveRecord::RecordInvalid.new(invalid_user))
 
-        post :sign_up
+        post :create
       end
 
       it 'has a 422 status code' do
@@ -133,7 +65,7 @@ RSpec.describe Api::V1::Auth::UsersController, type: :controller do
       before do
         allow(AuthService).to receive(:sign_up!).and_raise
 
-        post :sign_up
+        post :create
       end
 
       it 'has a 422 status code' do
@@ -141,39 +73,6 @@ RSpec.describe Api::V1::Auth::UsersController, type: :controller do
       end
 
       it 'returns json with errors' do
-        expect(response.body).to eq result.to_json
-      end
-    end
-  end
-
-  describe '#sign_out_user' do
-    context 'When user is signed' do
-      before do
-        allow(controller).to receive(:authenticate_request!).and_return(build_stubbed(:user))
-        post :sign_out_user
-      end
-
-      it 'has a 204 status code' do
-        expect(response.status).to eq(204)
-      end
-    end
-
-    context 'When user is not signed' do
-      let(:result) {
-        {
-          errors: ['Not Authenticated']
-        }
-      }
-
-      before do
-        post :sign_out_user
-      end
-
-      it 'has a 401 status code' do
-        expect(response.status).to eq(401)
-      end
-
-      it 'returns json with error' do
         expect(response.body).to eq result.to_json
       end
     end
